conda-store-server/pyproject.toml
[build-system]
requires = ["hatchling>=1.14.0"]
build-backend = "hatchling.build"

[project]
name = "conda-store-server"
description = "Conda Environment Management, Builds, and Serve"
readme = "README.md"
license = "BSD-3-Clause"
requires-python = ">=3.7"
keywords = [
  "conda",
]
authors = [
  { name = "Christopher Ostrouchov", email = "crhsi.ostrouchov@gmail.com" },
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Topic :: Software Development :: Build Tools",
  "License :: OSI Approved :: BSD License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3 :: Only",
]
dependencies = [
  "conda-docker",
  "conda-pack",
  "celery",
  "sqlalchemy",
  "requests",
  "fastapi",
  "ruamel.yaml",
  "redis",
  "pydantic<2",
  "minio",
  "traitlets",
  "pyjwt",
  "yarl",
  "filelock",
  "itsdangerous",
  "jinja2",
  "python-multipart",
  # conda (which should not be included here)
]
dynamic = ["version"]

[project.urls]
Homepage = "https://conda.store/"
Source = "https://github.com/conda-incubator/conda-store"

[project.optional-dependencies]
dev = [
  "build",
  "twine",
]

[tool.hatch.version]
path = "conda_store_server/__init__.py"

[tool.hatch.envs.dev]
dependencies = [
  "pre-commit",
  "sphinx",
  "myst-parser",
  "sphinx-panels",
  "sphinx-copybutton",
  "pydata-sphinx-theme",
  "pytest",
  "pytest-mock",
  "pytest-playwright",
]

[tool.hatch.envs.dev.scripts]
lint = [
  "pre-commit run --all",
]
docs = "sphinx-build -b html ../docs ../docs/_build"
unit-test = "pytest tests -v"
playwright-test = [
  "playwright install",
  "pytest --video on ../tests/test_playwright.py"
]
integration-test = [
  "pytest ../tests/test_api.py ../tests/test_metrics.py"
]

[tool.hatch.build.hooks.custom]

[tool.hatch.build.targets.sdist.hooks.custom]

[project.scripts]
conda-store-server = "conda_store_server.server.__main__:main"
conda-store-worker = "conda_store_server.worker.__main__:main"

[tool.black]
line-length = 88
target-version = ['py37', 'py38', 'py39']

[tool.ruff]
ignore = [
  "E501",  # line-length
]
